PluginCheck: resolve all EscapeOutput findings (65 -> 0)
Shipped-plugin PCP errors now 0 (the 2 PCP still reports are in
.distignore'd dev files: .wp-env.json, phpcs.xml.dist).

Two kinds of fix:

- Verified-safe output annotated with a justified phpcs:ignore:
  get_block_wrapper_attributes() (escaped HTML),
  wp_interactivity_data_wp_context() (escaped attribute), the
  sprintf-built $inner_attributes, static Petstablished_Icons SVG,
  pre-escaped $scoped_styles CSS, static $slide_directives, and
  rendered InnerBlocks $content.

- Genuine escaping added where output was a raw scalar: (int) casts on
  counts/totals/durations (comparison count, max slots, batch_size,
  fav_count, active filters, result totals), esc_html() on the
  search-result _n() and fav_count, and esc_js() on the admin sync
  nonces echoed into inline JS.

Verified: archive, single pet, comparison, and search all render 200
with no errors; counts display correctly.

+++ blocks/pet-compare-bar/render.php
@@ -100,7 +100,7 @@ $wrapper_attributes = get_block_wrapper_attributes( array(
 ) );
 ?>
 
-<aside <?php echo $wrapper_attributes; ?>>
+<aside <?php echo $wrapper_attributes; /* phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- get_block_wrapper_attributes() returns escaped HTML. */ ?>>
 	<!-- Collapsed pill — visible when bar is collapsed -->
 	<button
 		type="button"
@@ -113,7 +113,7 @@ $wrapper_attributes = get_block_wrapper_attributes( array(
 		<span
 			class="pet-compare-bar__pill-count"
 			data-wp-text="petstablished::state.comparisonCount"
-		><?php echo count( $comparison ); ?></span>
+		><?php echo (int) count( $comparison ); ?></span>
 	</button>
 
 	<!-- Expanded bar — the full interface -->
@@ -121,7 +121,7 @@ $wrapper_attributes = get_block_wrapper_attributes( array(
 		<div class="pet-compare-bar__header">
 			<span class="pet-compare-bar__label">
 				<?php esc_html_e( 'Compare', 'vcpahumane-pet-sync' ); ?>
-				(<span data-wp-text="petstablished::state.comparisonCount"><?php echo count( $comparison ); ?></span>/<?php echo $max_compare; ?>)
+				(<span data-wp-text="petstablished::state.comparisonCount"><?php echo (int) count( $comparison ); ?></span>/<?php echo (int) $max_compare; ?>)
 			</span>
 			<button
 				type="button"
@@ -155,7 +155,7 @@ $wrapper_attributes = get_block_wrapper_attributes( array(
 			?>
 				<div
 					class="pet-compare-bar__slot"
-					<?php echo wp_interactivity_data_wp_context( $slot_context ); ?>
+					<?php echo wp_interactivity_data_wp_context( $slot_context ); /* phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- wp_interactivity_data_wp_context() returns an escaped attribute. */ ?>
 					data-wp-class--has-pet="state.slotHasPet"
 				>
 					<div
